src/third_party/blink/web_tests/serial/serialPort_clearReadError.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/mojo/public/mojom/base/unguessable_token.mojom.js"></script>
<script src="file:///gen/third_party/blink/public/mojom/serial/serial.mojom.js"></script>
<script src="resources/serial-test-utils.js"></script>
<script>

serial_test(async (t, fake) => {
  const { port, fakePort } = await getFakeSerialPort(fake);

  assert_throws('InvalidStateError', () => port.clearReadError());
}, 'clearReadError() not allowed on a closed port.');

serial_test(async (t, fake) => {
  const { port, fakePort } = await getFakeSerialPort(fake);

  await port.open({ baudrate: 9600 });
  assert_throws('InvalidStateError', () => port.clearReadError());
}, 'clearReadError() not allowed without a reported error');

serial_test(async (t, fake) => {
  const { port, fakePort } = await getFakeSerialPort(fake);
  await port.open({ baudrate: 9600 });

  fakePort.simulateParityError();

  assert_throws('InvalidStateError', () => port.clearReadError());
}, 'clearReadError() not allowed before readable has been aborted');

serial_test(async (t, fake) => {
  const { port, fakePort } = await getFakeSerialPort(fake);
  await port.open({ baudrate: 9600 });

  fakePort.simulateParityError();

  await promise_rejects(t, 'NetworkError', port.readable.getReader().read());
  port.clearReadError();
}, 'clearReadError() can be called after readable reports error');

</script>
